+++ src/S3.php
@@ -2,8 +2,10 @@
 
 namespace Sfneal\Helpers\Aws\S3;
 
+use Closure;
 use DateTimeInterface;
 use Illuminate\Contracts\Filesystem\FileNotFoundException;
+use Illuminate\Contracts\Filesystem\Filesystem;
 use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\Storage;
 
@@ -31,6 +33,16 @@ class S3
     }
 
     /**
+     * Retrieve a Filesystem instance for the specified disk.
+     *
+     * @return Filesystem
+     */
+    private function storageDisk(): Filesystem
+    {
+        return Storage::disk($this->disk);
+    }
+
+    /**
      * Set the filesystem disk.
      *
      * @param string $disk
@@ -53,9 +65,9 @@ class S3
     public function url(bool $temp = true, DateTimeInterface $expiration = null): string
     {
         if ($temp) {
-            return Storage::disk($this->disk)->temporaryUrl($this->s3_key, $expiration ?? now()->addMinutes(60));
+            return $this->storageDisk()->temporaryUrl($this->s3_key, $expiration ?? now()->addMinutes(60));
         } else {
-            return Storage::disk($this->disk)->url($this->s3_key);
+            return $this->storageDisk()->url($this->s3_key);
         }
     }
 
@@ -77,7 +89,7 @@ class S3
      */
     public function exists(): bool
     {
-        return Storage::disk($this->disk)->exists($this->s3_key);
+        return $this->storageDisk()->exists($this->s3_key);
     }
 
     /**
@@ -90,9 +102,9 @@ class S3
     public function upload(string $file_path, string $acl = null): string
     {
         if (is_null($acl)) {
-            Storage::disk($this->disk)->put($this->s3_key, fopen($file_path, 'r+'));
+            $this->storageDisk()->put($this->s3_key, fopen($file_path, 'r+'));
         } else {
-            Storage::disk($this->disk)->put($this->s3_key, fopen($file_path, 'r+'), $acl);
+            $this->storageDisk()->put($this->s3_key, fopen($file_path, 'r+'), $acl);
         }
 
         return $this->url();
@@ -108,9 +120,9 @@ class S3
     public function upload_raw(string $file_contents, string $acl = null): string
     {
         if (is_null($acl)) {
-            Storage::disk($this->disk)->put($this->s3_key, $file_contents);
+            $this->storageDisk()->put($this->s3_key, $file_contents);
         } else {
-            Storage::disk($this->disk)->put($this->s3_key, $file_contents, $acl);
+            $this->storageDisk()->put($this->s3_key, $file_contents, $acl);
         }
 
         return $this->url();
@@ -129,8 +141,8 @@ class S3
             $file_name = basename($this->s3_key);
         }
 
-        $mime = Storage::disk($this->disk)->getMimetype($this->s3_key);
-        $size = Storage::disk($this->disk)->getSize($this->s3_key);
+        $mime = $this->storageDisk()->getMimetype($this->s3_key);
+        $size = $this->storageDisk()->getSize($this->s3_key);
 
         $response = [
             'Content-Type' => $mime,
@@ -140,7 +152,7 @@ class S3
             'Content-Transfer-Encoding' => 'binary',
         ];
 
-        return Response::make(Storage::disk($this->disk)->get($this->s3_key), 200, $response);
+        return Response::make($this->storageDisk()->get($this->s3_key), 200, $response);
     }
 
     /**
@@ -150,7 +162,7 @@ class S3
      */
     public function delete(): bool
     {
-        return Storage::disk($this->disk)->delete($this->s3_key);
+        return $this->storageDisk()->delete($this->s3_key);
     }
 
     /**
@@ -160,7 +172,7 @@ class S3
      */
     public function list(): array
     {
-        $storage = Storage::disk($this->disk);
+        $storage = $this->storageDisk();
         $client = $storage->getAdapter()->getClient();
         $command = $client->getCommand('ListObjects');
         $command['Bucket'] = $storage->getAdapter()->getBucket();
@@ -181,5 +193,51 @@ class S3
         }
 
         return $files;
+    }
+
+    /**
+     * Autocomplete an S3 path by providing the known start of a path.
+     *
+     * - once path autocompletion is resolved the $s3_key property is replaced with the found path
+     *
+     * @return $this
+     */
+    public function autocompletePath(): self
+    {
+        // Extract the known $base of the path & the $wildcard
+        $base = dirname($this->s3_key);
+        $wildcard = basename($this->s3_key);
+
+        // Get all of the folders in the base directory
+        $folders = $this->storageDisk()->directories($base);
+
+        // Filter folders to find the wildcard path
+        $folders = array_filter($folders, function ($value) use ($base, $wildcard) {
+            return str_starts_with($value, $base.DIRECTORY_SEPARATOR.$wildcard);
+        });
+
+        // return the resolved path
+        $this->s3_key = collect($folders)->values()->first();
+
+        return $this;
+    }
+
+    /**
+     * Retrieve an array of all files in a directory with an optional filtering closure.
+     *
+     * @param Closure|null $closure
+     * @return array
+     */
+    public function allFiles(Closure $closure = null): array
+    {
+        // Create array of all files
+        $allFiles = $this->storageDisk()->allFiles($this->s3_key);
+
+        // Apply filtering closure
+        if (isset($closure)) {
+            return array_filter(array_values($allFiles), $closure);
+        }
+
+        return $allFiles;
     }
 }